Changed Data to TimerPrinter and moved it to a new file
